# backend/api/emotion/router.py
"""
Emotion API Router

REST API endpoints for AICO's emotion system.
Provides access to current emotional state and history.
"""

from typing import Annotated, Optional
from datetime import datetime, timedelta, UTC
from fastapi import APIRouter, Depends, Query
import os
from aico.core.logging import get_logger
from .schemas import EmotionStateResponse, EmotionHistoryResponse, EmotionHistoryItem
from .dependencies import get_current_user, get_emotion_engine
from backend.core.postgres_dependencies import get_uow
from aico.data.uow import UnitOfWork
from backend.api.errors import raise_api_error

logger = get_logger("aico.api.emotion.router")

router = APIRouter()


@router.get("/current", response_model=EmotionStateResponse)
async def get_current_emotion(
    user: Annotated[dict, Depends(get_current_user)]
):
    """
    Get AICO's current emotional state.
    
    **Authentication required:** Bearer token
    
    Returns the most recent emotional state generated by the emotion simulation engine,
    including primary emotion label, confidence, and dimensional values (valence, arousal, dominance).
    
    **Note:** Current implementation returns global emotional state. Future versions will support
    per-user emotional states for personalized interactions.
    """
    try:
        from backend.api_gateway.core.nats_client import get_gateway_nats_client
        
        # Request current emotion from core via NATS
        nats_client = get_gateway_nats_client()
        emotion_data = await nats_client.request_current_emotion()
        
        return EmotionStateResponse(**emotion_data)
        
    except Exception as e:
        logger.error(f"Error retrieving current emotion via NATS: {e}")
        from fastapi.responses import JSONResponse
        return JSONResponse(
            status_code=500,
            content={
                "error_code": "EMOTION_STATE_FETCH_FAILED",
                "message": "Failed to retrieve emotional state"
            }
        )


@router.get("/history", response_model=EmotionHistoryResponse)
async def get_emotion_history(
    user: Annotated[dict, Depends(get_current_user)],
    limit: int = Query(50, ge=1, le=1000, description="Maximum number of records to return"),
    hours: Optional[int] = Query(None, ge=1, description="Only return emotions from last N hours"),
    days: Optional[int] = Query(None, ge=1, description="Only return emotions from last N days"),
    since: Optional[str] = Query(None, description="ISO timestamp - only return emotions after this time"),
    feeling: Optional[str] = Query(None, description="Filter by specific emotion (e.g., 'calm', 'playful')")
):
    """
    Get AICO's emotional state history from database.
    
    **Authentication required:** Bearer token
    
    Returns a list of emotional states for mood arc visualization, queried directly from the database.
    States are ordered from most recent to oldest.
    
    **Smart filtering options:**
    - `limit`: Max records (1-1000, default: 50)
    - `hours`: Last N hours (e.g., hours=24 for last day)
    - `days`: Last N days (e.g., days=7 for last week)
    - `since`: ISO timestamp (e.g., 2025-11-20T00:00:00)
    - `feeling`: Filter by emotion (e.g., 'calm', 'playful', 'warm_concern')
    
    **Examples:**
    - `/history?limit=100` - Last 100 emotions
    - `/history?days=7` - All emotions from last 7 days
    - `/history?hours=24&feeling=calm` - All 'calm' emotions in last 24 hours
    - `/history?since=2025-11-20T00:00:00` - All emotions since Nov 20
    
    Args:
        user: Authenticated user (injected)
        emotion_engine: Emotion engine service (injected)
        limit: Maximum number of history entries to return
        hours: Filter to last N hours
        days: Filter to last N days
        since: Filter to emotions after this timestamp
        feeling: Filter by specific emotion label
    """
    try:
        from backend.api_gateway.core.nats_client import get_gateway_nats_client
        
        # Calculate hours from days if provided
        if days:
            hours = days * 24
        elif not hours:
            hours = 24
        
        # Request emotion history from core via NATS
        nats_client = get_gateway_nats_client()
        history_data = await nats_client.request_emotion_history(
            limit=limit,
            hours=hours
        )
        
        return EmotionHistoryResponse(**history_data)
    except Exception as e:
        logger.error(f"Error retrieving emotion history via NATS: {e}")
        # Return JSONResponse directly to avoid double-response issue with encryption middleware
        from fastapi.responses import JSONResponse
        return JSONResponse(
            status_code=503,
            content={
                "error_code": "EMOTION_ENGINE_UNAVAILABLE",
                "message": "Emotion engine unavailable"
            }
        )