# openapi/public-openapi.json
{
  "components": {
    "schemas": {
...
      "EmotionHistoryItem": {
        "description": "Single emotional state history entry",
        "example": {
          "arousal": 0.5,
          "feeling": "warm_concern",
          "intensity": 0.7,
          "timestamp": "2024-11-19T15:30:00Z",
          "valence": 0.3
        },
        "properties": {
          "arousal": {
            "description": "Mood arousal",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Arousal",
            "type": "number"
          },
          "feeling": {
            "description": "Subjective feeling label",
            "title": "Feeling",
            "type": "string"
          },
          "intensity": {
            "description": "Emotional intensity",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Intensity",
            "type": "number"
          },
          "timestamp": {
            "description": "ISO 8601 timestamp",
            "title": "Timestamp",
            "type": "string"
          },
          "valence": {
            "description": "Mood valence",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Valence",
            "type": "number"
          }
        },
        "required": [
          "timestamp",
          "feeling",
          "valence",
          "arousal",
          "intensity"
        ],
        "title": "EmotionHistoryItem",
        "type": "object"
      },
      "EmotionHistoryResponse": {
        "description": "Emotional state history response",
        "example": {
          "count": 2,
          "history": [
            {
              "arousal": 0.5,
              "feeling": "warm_concern",
              "intensity": 0.7,
              "timestamp": "2024-11-19T15:30:00Z",
              "valence": 0.3
            },
            {
              "arousal": 0.5,
              "feeling": "neutral",
              "intensity": 0.5,
              "timestamp": "2024-11-19T15:29:00Z",
              "valence": 0.0
            }
          ]
        },
        "properties": {
          "count": {
            "description": "Number of history entries",
            "title": "Count",
            "type": "integer"
          },
          "history": {
            "description": "List of emotional states",
            "items": {
              "$ref": "#/components/schemas/EmotionHistoryItem"
            },
            "title": "History",
            "type": "array"
          }
        },
        "required": [
          "count",
          "history"
        ],
        "title": "EmotionHistoryResponse",
        "type": "object"
      },
      "EmotionStateResponse": {
        "description": "Current emotional state response",
        "example": {
          "arousal": 0.4,
          "confidence": 0.85,
          "dominance": 0.5,
          "primary": "calm",
          "timestamp": "2024-11-19T15:30:00Z",
          "valence": 0.3
        },
        "properties": {
          "arousal": {
            "description": "Activation/energy level",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Arousal",
            "type": "number"
          },
          "confidence": {
            "description": "Confidence level",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "dominance": {
            "description": "Control/power dimension",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Dominance",
            "type": "number"
          },
          "primary": {
            "description": "Primary emotion label",
            "title": "Primary",
            "type": "string"
          },
          "timestamp": {
            "description": "ISO 8601 timestamp of emotional state",
            "title": "Timestamp",
            "type": "string"
          },
          "valence": {
            "description": "Pleasure/displeasure dimension",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Valence",
            "type": "number"
          }
        },
        "required": [
          "timestamp",
          "primary",
          "confidence",
          "valence",
          "arousal",
          "dominance"
        ],
        "title": "EmotionStateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AICO Backend REST API with clean architecture",
    "title": "AICO Backend API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/emotion/current": {
      "get": {
        "description": "Get AICO's current emotional state.\n\n**Authentication required:** Bearer token\n\nReturns the most recent emotional state generated by the emotion simulation engine,\nincluding primary emotion label, confidence, and dimensional values (valence, arousal, dominance).\n\n**Note:** Current implementation returns global emotional state. Future versions will support\nper-user emotional states for personalized interactions.",
        "operationId": "get_current_emotion_api_v1_emotion_current_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmotionStateResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Current Emotion",
        "tags": [
          "emotion"
        ]
      }
    },
    "/api/v1/emotion/history": {
      "get": {
        "description": "Get AICO's emotional state history from database.\n\n**Authentication required:** Bearer token\n\nReturns a list of emotional states for mood arc visualization, queried directly from the database.\nStates are ordered from most recent to oldest.\n\n**Smart filtering options:**\n- `limit`: Max records (1-1000, default: 50)\n- `hours`: Last N hours (e.g., hours=24 for last day)\n- `days`: Last N days (e.g., days=7 for last week)\n- `since`: ISO timestamp (e.g., 2025-11-20T00:00:00)\n- `feeling`: Filter by emotion (e.g., 'calm', 'playful', 'warm_concern')\n\n**Examples:**\n- `/history?limit=100` - Last 100 emotions\n- `/history?days=7` - All emotions from last 7 days\n- `/history?hours=24&feeling=calm` - All 'calm' emotions in last 24 hours\n- `/history?since=2025-11-20T00:00:00` - All emotions since Nov 20\n\nArgs:\n    user: Authenticated user (injected)\n    emotion_engine: Emotion engine service (injected)\n    limit: Maximum number of history entries to return\n    hours: Filter to last N hours\n    days: Filter to last N days\n    since: Filter to emotions after this timestamp\n    feeling: Filter by specific emotion label",
        "operationId": "get_emotion_history_api_v1_emotion_history_get",
        "parameters": [
          {
            "description": "Maximum number of records to return",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Maximum number of records to return",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Only return emotions from last N hours",
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only return emotions from last N hours",
              "title": "Hours"
            }
          },
          {
            "description": "Only return emotions from last N days",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only return emotions from last N days",
              "title": "Days"
            }
          },
          {
            "description": "ISO timestamp - only return emotions after this time",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO timestamp - only return emotions after this time",
              "title": "Since"
            }
          },
          {
            "description": "Filter by specific emotion (e.g., 'calm', 'playful')",
            "in": "query",
            "name": "feeling",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by specific emotion (e.g., 'calm', 'playful')",
              "title": "Feeling"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmotionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Emotion History",
        "tags": [
          "emotion"
        ]
      }
    },
...
  }
}